improve auto refreshing of results
